Move 'cookies' property from RequestHandler to HTTPRequest
